# Support no validation to improve write performance