@geky 2020-02-24
This is interesting, thanks for creating a pr!

I haven't considered a non-validation option a priority since usually the time spent erasing greatly outweighs the overhead of bus transactions. But I don't see a reason to not bring this in. It may also have more benefits for a multi-tasking system that can perform other operations during the erase time.

I was hurt before working on a project with many configuration options, as it made testing every configuration a pain. Though maybe littlefs is mature enough that it should expand the available configurations?

Currently https://github.com/ARMmbed/littlefs/pull/372 and related bug fixes is my priority, so I may wait on this until after the next release. Sorry about the delay.

---

It would also be interesting to look into improvements on how littlefs is configured in the same release https://github.com/ARMmbed/littlefs/issues/158#issuecomment-494891516 (though this should be a different PR)

> At first glance it seems like you could balance the compile time and runtime concerns with something like this in lfs_utils.h
>
> #define LFS_ALIGNMENT(lfs) lfs->config->cache_alignment
>
> and similar for many other configurables.
>
> You wouldn't be able to entirely get rid of the runtime cost using this mechanism, but for people who don't need these features, replacing the macros to return constants would potentially allow constant folding to collapse some of the work.