@geky 2020-02-24
There's actually two different methods littlefs uses to check disk-writes. You found the one used by files, but a separate method is used in lfs_dir_commitcrc to check dir commits without need to store the data in RAM. 

You can probably skip this loop as well:
https://github.com/ARMmbed/littlefs/blob/88bc3335599bc4d4655cde38b4965099ec28ce00/lfs.c##L1297