@gmpy 2020-02-26
> There's actually two different methods littlefs uses to check disk-writes. You found the one used by files, but a separate method is used in lfs_dir_commitcrc to check dir commits without need to store the data in RAM.
> 
> You can probably skip this loop as well:
> https://github.com/ARMmbed/littlefs/blob/88bc3335599bc4d4655cde38b4965099ec28ce00/lfs.c##L1297

I did miss it. But it has little impact on write performance. On my test on a 10M file, lfs_dir_commitcrc() is called twice and lfs_bd_flush() at lfs_bd_prog() is called 2565 times.

The performance as follow:

	read average speed: 12362.17 KB/s
	write average speed: 2691.20 KB/s

I suggest that keep checking here, which makes the cache of dir on flash drivers keeping hot since dir is read frequently. It pays little performance to check dir.