@pjsg 2020-02-26
One important feature of the current verification is that it catches logical errors where the programmed data is not written over 'erased' memory, but over 'used' memory. It would be a pity to trade off detection of corruption for a bit of performance.