@gmpy 2020-02-28
> One important feature of the current verification is that it catches logical errors where the programmed data is not written over 'erased' memory, but over 'used' memory. It would be a pity to trade off detection of corruption for a bit of performance.

I think 30% increase has been a lot. This is very important for some products with performance requirements. This is a switch, the user can choose whether to turn it on or not. To me, it can be enabled during the development phase, and can be closed after the product is formed for the highest experience.