@pjsg 2020-03-04
You might want to check that (at least) the first byte of the area that you are about to program is set to the erase value. If it isn't, then you should abort.