@gmpy 2020-03-09
> You might want to check that (at least) the first byte of the area that you are about to program is set to the erase value. If it isn't, then you should abort.

The minimum read size for nand/mmc is equal to page size. Take my using spinand as am example, the page size is 2KB. So reading one byte may be no difference to reading a page.

In addition, during my analysis, the nand flash driver always has ram cache. The lfs writes to cache on driver and reads for validation from cache. It should not be much different. There may be some unexpected situations. Please put this patch aside before i find it out.